+++ 2-mod/simplest_ot.cu
@@ -27,19 +27,19 @@ SimplestOT::~SimplestOT() {
     simplestOTRecvers[id] = nullptr;
 }
 
-void SimplestOT::fromOwnBuffer(uint8_t *d, size_t nBytes) {
+void SimplestOT::fromOwnBuffer(uint8_t *d, uint64_t nBytes) {
   while (!hasContent);
   memcpy(d, buffer, nBytes);
   hasContent = false;
 }
 
-void SimplestOT::toOtherBuffer(uint8_t *s, size_t nBytes) {
+void SimplestOT::toOtherBuffer(uint8_t *s, uint64_t nBytes) {
   while (other->hasContent);
   memcpy(other->buffer, s, nBytes);
   other->hasContent = true;
 }
 
-std::array<std::vector<GPUBlock>, 2> SimplestOT::send(size_t count) {
+std::array<std::vector<GPUBlock>, 2> SimplestOT::send(uint64_t count) {
   a.randomize(prng);
   A = Point::mulGenerator(a);
   toOtherBuffer((uint8_t*) &A, sizeof(A));
@@ -51,7 +51,7 @@ std::array<std::vector<GPUBlock>, 2> SimplestOT::send(size_t count) {
   A *= a;
   fromOwnBuffer((uint8_t*) &B.at(0), sizeof(B.at(0)) * B.size());
 
-  for (size_t i = 0; i < count; i++) {
+  for (uint64_t i = 0; i < count; i++) {
     B.at(i) *= a;
     osuCrypto::RandomOracle ro(BLK_SIZE);
     ro.Update(B.at(i));
@@ -70,10 +70,10 @@ std::array<std::vector<GPUBlock>, 2> SimplestOT::send(size_t count) {
   return m;
 }
 
-std::vector<GPUBlock> SimplestOT::recv(size_t count, uint64_t choice) {
+std::vector<GPUBlock> SimplestOT::recv(uint64_t count, uint64_t choice) {
   fromOwnBuffer((uint8_t*) &A, sizeof(A));
   std::vector<GPUBlock> mb(count, BLK_SIZE);
-  for (size_t i = 0; i < count; i++) {
+  for (uint64_t i = 0; i < count; i++) {
     b.emplace_back(prng);
     Point B0 = Point::mulGenerator(b.at(i));
     Point B1 = A + B0;
@@ -83,7 +83,7 @@ std::vector<GPUBlock> SimplestOT::recv(size_t count, uint64_t choice) {
   toOtherBuffer((uint8_t*) &B.at(0), sizeof(B.at(0)) * B.size());
 
   uint8_t buff[BLK_SIZE];
-  for (size_t i = 0; i < count; i++) {
+  for (uint64_t i = 0; i < count; i++) {
     Point mB = A * b.at(i);
     osuCrypto::RandomOracle ro(BLK_SIZE);
     ro.Update(mB);
