+++ 3-dev/aes_expand.h
@@ -6,6 +6,6 @@
 
 __global__
 void aesExpand128(unsigned *aesKey, TreeNode *leaves, uint32_t *m,
-	unsigned *inData, int expandDir, size_t width);
+	unsigned *inData, int expandDir, uint64_t width);
 
 #endif
