+++ 3-dev/basic_op.cu
@@ -1,22 +1,22 @@
 #include "basic_op.h"
 
 __global__
-void xor_gpu(uint8_t *c, uint8_t *a, uint8_t *b, size_t n) {
-  size_t x = blockIdx.x * blockDim.x + threadIdx.x;
+void xor_gpu(uint8_t *c, uint8_t *a, uint8_t *b, uint64_t n) {
+  uint64_t x = blockIdx.x * blockDim.x + threadIdx.x;
   if (x < n)
     c[x] = a[x] ^ b[x];
 }
 
 __global__
-void xor_circular(uint8_t *c, uint8_t *a, uint8_t *b, size_t len_b, size_t n) {
+void xor_circular(uint8_t *c, uint8_t *a, uint8_t *b, uint64_t len_b, uint64_t n) {
   int x = blockIdx.x * blockDim.x + threadIdx.x;
   if (x < n)
     c[x] = a[x] ^ b[x % len_b];
 }
 
 __global__
-void and_gpu(uint8_t *c, uint8_t *a, size_t n) {
-  size_t x = blockIdx.x * blockDim.x + threadIdx.x;
+void and_gpu(uint8_t *c, uint8_t *a, uint64_t n) {
+  uint64_t x = blockIdx.x * blockDim.x + threadIdx.x;
   if (x < n)
     c[x] &= a[x];
 }
@@ -49,19 +49,19 @@ void chinese_rem_theorem_gpu(uint32_t *c, uint32_t d, uint32_t p,
 
 __global__
 void sum_reduce(uint64_t *c) {
-  size_t numThreads = gridDim.x * blockDim.x;
-  size_t destId = blockIdx.x * blockDim.x + threadIdx.x;
-  size_t srcId = destId + numThreads;
+  uint64_t numThreads = gridDim.x * blockDim.x;
+  uint64_t destId = blockIdx.x * blockDim.x + threadIdx.x;
+  uint64_t srcId = destId + numThreads;
   c[destId] ^= c[srcId];
 }
 
 // https://developer.download.nvidia.com/assets/cuda/files/reduction.pdf
 __global__
-void sum_gpu(uint64_t *c, size_t n) {
+void sum_gpu(uint64_t *c, uint64_t n) {
   extern __shared__ uint64_t sdata[];
-  size_t tid = threadIdx.x;
-  size_t i = blockIdx.x * blockDim.x * 2 + threadIdx.x;
-  size_t gridSize = blockDim.x * 2 * gridDim.x;
+  uint64_t tid = threadIdx.x;
+  uint64_t i = blockIdx.x * blockDim.x * 2 + threadIdx.x;
+  uint64_t gridSize = blockDim.x * 2 * gridDim.x;
   sdata[tid] = 0;
 
   for(; i < n; i += gridSize) {
