+++ 3-dev/rand.h
@@ -4,7 +4,7 @@
 #include "util.h"
 #include <curand_kernel.h>
 
-Matrix init_rand(curandGenerator_t &prng, size_t height, size_t width);
+Matrix init_rand(curandGenerator_t &prng, uint64_t height, uint64_t width);
 void gen_rand(curandGenerator_t prng, Matrix randMatrix);
 void del_rand(curandGenerator_t prng, Matrix randMatrix);
